+++ changelog.md
@@ -11,6 +11,8 @@ Codion Change Log
 - BaseValue extracted from AbstractValue, AbstractValueCollection extends it, now properly synchronized.
 ### is.codion.common.model
 - JsonPreferencesStore bug fixed, lock file now cleanup up on save.
+### is.codion.framework.model
+- EntityApplicationModel.version() removed.
 ### is.codion.swing.framework.ui
 - EntityTablePanelPreferences bug fixed, column width preferences got lost for panels that were never initialized, now sets the column width as well as preferred width when applying preferences.
 - EntityTablePanelPreferences, auto-resize-mode preference added.